+++ PKG-INFO
@@ -1,5 +1,5 @@
 Metadata-Version: 1.1
-Name: trolly
+Name: jirate
 Version: 0.4.0
 Summary: Python CLI for JIRA and Trello
 Author: Lon Hohberger
