+++ jirate/config.py
@@ -5,7 +5,7 @@ import os
 
 
 def get_config():
-    config_file = open(os.path.expanduser('~/.trolly.json'))
+    config_file = open(os.path.expanduser('~/.jirate.json'))
     config_data = config_file.read()
     config_file.close()
     config = json.loads(config_data)
