+++ templates/template.yaml
@@ -1,6 +1,6 @@
 issues:
   - issue_type: "Story"
-    summary: "Set up Jolly templates on my machine"
+    summary: "Set up Jirate templates on my machine"
     # Description terminated by two newlines
     description: |
       * {*}This is a multi-line description.{*}
